requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==5.0.3
Flask==1.1.2
Flask-Cors==3.0.10
geographiclib==1.50
geopy==2.1.0
gunicorn==20.1.0
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
MarkupSafe==1.1.1
osm2geojson==0.1.29
overpy==0.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
python-dateutil==2.8.1
python-dotenv==0.16.0
pyzmq==22.0.3
requests==2.25.1
Shapely==1.7.1
six==1.15.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
